upcoming events list page, events endpoint clean up, event create endpoint clean up

+++ api/controllers/events.js
@@ -1,8 +1,10 @@
 const { GetObjectCommand, PutObjectCommand } = require('@aws-sdk/client-s3');
 const fs = require('fs');
 const { getSignedUrl } = require("@aws-sdk/s3-request-presigner");
+const jwt = require('jsonwebtoken');
+const { Op } = require("sequelize");
 
-const { Events } = require('../models');
+const { Events, Users } = require('../models');
 const { isAllowedFileType } = require('../helpers');
 const { s3Client } = require('../clients/s3');
 const { awsRegion } = require('../static');
@@ -11,34 +13,42 @@ const { awsRegion } = require('../static');
 const Bucket = process.env.S3_BUCKET_NAME;
 
 const createEvent = async (req, res) => {
-    
-    // get user and make sure they're admin
-    // get user and make sure they're admin
-    // get user and make sure they're admin
+    const cookie = req.signedCookies.hackr_harvest;
+    const { id: userId } = jwt.verify(cookie, process.env.JWT_SECRET);
+    const { dataValues: user } = await Users.findOne({ where: { id: userId } });
 
-    if (!isAllowedFileType(req.file.mimetype)) {
-        return res.status(400).send('event image file type not allowed');
+    if (user.role !== 'admin') {
+        return res.status(403).send('not authorized');
     }
 
-    const file = fs.readFileSync(req.file.path);
-    const Key = `${req.body.name}-eventImg`;
-    const command = new PutObjectCommand({
-        Bucket,
-        Key,
-        Body: file,
-        ContentType: req.file.mimetype,
-    });
-    await s3Client.send(command);
-    fs.unlink(req.file.path, () => {});
+    let imageUrl = null;
+    let signed = null;
+    if (req.file) {
+        if (!isAllowedFileType(req.file.mimetype)) {
+            return res.status(400).send('event image file type not allowed');
+        }
+
+        const file = fs.readFileSync(req.file.path);
+        const Key = `${req.body.name}-eventImg`;
+        const command = new PutObjectCommand({
+            Bucket,
+            Key,
+            Body: file,
+            ContentType: req.file.mimetype,
+        });
+        await s3Client.send(command);
+        fs.unlink(req.file.path, () => {});
+        imageUrl = `https://${Bucket}.s3.${awsRegion}.amazonaws.com/${Key}`;
+        const command2 = new GetObjectCommand({ Bucket, Key });
+        signed = await getSignedUrl(s3Client, command2, { expiresIn: 3600 });
+    }
 
     const event = await Events.create({
         ...req.body,
-        hostId: 1,
-        imageUrl: `https://${Bucket}.s3.${awsRegion}.amazonaws.com/${Key}`,
+        hostId: userId,
+        imageUrl,
     });
 
-    const command2 = new GetObjectCommand({ Bucket, Key });
-    const signed = await getSignedUrl(s3Client, command2, { expiresIn: 3600 });
     event.imageUrl = signed;
 
     res.send({ event });
@@ -65,13 +75,14 @@ const getEventById = async (req, res) => {
 };
 
 const getAllEvents = async (req, res) => {
-    const { limit = 10, offset = 0 } = req.query;
+    const { limit = 10, offset = 0, past = false } = req.query;
     const events = await Events.findAll({
         include: ['host', 'attendees'],
         order: [
             ['startTime', 'DESC'],
             ['name', 'ASC'],
         ],
+        where: { startTime: { [past ? Op.lt : Op.gt]: new Date(Date.now()) } },
         limit,
         offset,
     });
